requirements.txt
flask
sqlalchemy
flask-cors
google-generativeai
python-dotenv
